Split engine local copy tests into dedicated module
